code/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vision Language Model
qwen-vl-utils>=0.0.1

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
# Alternative: pytesseract>=0.3.10

# RAG & Retrieval
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4  # or faiss-gpu if using GPU
wikipedia>=1.4.0

# Vietnamese NLP
underthesea>=1.3.0

# Gradio for demo
gradio>=4.0.0

# Utilities
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# For notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

